Fixed bugs with checkpointing and 'all-linear' type training

+++ src/model/model.py
@@ -45,10 +45,14 @@ class BaseMixin(AbstractMixin):
         if not os.path.exists(output_dir):
             os.makedirs(output_dir)
         self.output_path = os.path.join(output_dir, "dynalora_logs.json")
-        data = {"schedule": self.peft_config["dynalora"].schedule_type, "allocator": self.peft_config["dynalora"].allocator_type, "aggregate": self.peft_config["dynalora"].aggregate_type,
-                "adapter_base_names":self.adapter_base_names, "cum_acts": [], "masks": []}
-        with open(self.output_path, "w") as f:
-            json.dump(data, f)
+        if not os.path.exists(self.output_path):
+            data = {"schedule": self.peft_config[self.adapter_name].schedule_type,
+                    "allocator": self.peft_config[self.adapter_name].allocator_type,
+                    "aggregate": self.peft_config[self.adapter_name].aggregate_type,
+                    "adapter_base_names": list(self.named_adapter_modules.keys()),
+                    "cum_acts": [], "masks": []}
+            with open(self.output_path, "w") as f:
+                json.dump(data, f)
 
     @classmethod
     def _create_new_module(cls, lora_config, adapter_name, target, **kwargs):
@@ -84,10 +88,10 @@ class BaseMixin(AbstractMixin):
         """
 
         if not hasattr(self, "adapter_modules"):
-            self.adapter_modules = self._find_adapter_modules()
+            self.named_adapter_modules = self._find_adapter_modules()
         # use the *configured* choice function to obtain the k modules we want to activate
-        mask = self.allocator([1 for mod in self.adapter_modules])
-        for mod, msk in zip(self.adapter_modules, mask):
+        self.mask = self.allocator([1 for mod in self.named_adapter_modules.values()])
+        for mod, msk in zip(self.named_adapter_modules.values(), self.mask):
             if msk:
                 mod.activate()
             else:
@@ -103,42 +107,50 @@ class BaseMixin(AbstractMixin):
             )
 
         if not hasattr(self, "adapter_modules"):
-            self.adapter_modules = self._find_adapter_modules()
+            self.named_adapter_modules = self._find_adapter_modules()
 
         # use the *configured* choice function to obtain the k modules we want to activate
-        cum_acts = [mod.cum_acts for mod in self.adapter_modules]
-        mask = self.allocator(cum_acts)
+        cum_acts = [mod.cum_acts for mod in self.named_adapter_modules.values()]
+        self.mask = self.allocator(cum_acts)
 
         # activate the k modules, deactivate the rest.
         #   this is easiest via linear search and
         #   O(1) lookup
-        for mod, msk in zip(self.adapter_modules, mask):
+        for mod, msk in zip(self.named_adapter_modules.values(), self.mask):
             if msk:
                 mod.activate()
             else:
                 mod.deactivate()
-                
+
         # log to json
         with open(self.output_path, "r") as f:
             data = json.load(f)
             data["cum_acts"].append([cum_act.item() for cum_act in cum_acts])
-            data["masks"].append(mask.tolist())
+            data["masks"].append(self.mask.tolist())
         with open(self.output_path, "w") as f:
             json.dump(data, f)
-            
 
     def _find_adapter_modules(self):
         """
         Find all adapter modules in the model
         """
-        adapter_modules = []
-        self.adapter_base_names = []
+        named_adapter_modules = {}
         for name, module in self.model.named_modules():
             if isinstance(module, self.applicable_modules):
-                adapter_modules.append(module)
                 # this will be logged to json
-                self.adapter_base_names.append(name)
-        return adapter_modules
+                named_adapter_modules[name] = module
+        return named_adapter_modules
+
+    def set_mask(self, mask):
+        """
+        Set the mask of the model
+        """
+        self.mask = mask
+        for mod, msk in zip(self.named_adapter_modules.values(), self.mask):
+            if msk:
+                mod.activate()
+            else:
+                mod.deactivate()
 
 # DynaLoRA
 class DynaLoraMixin(BaseMixin):
@@ -155,17 +167,7 @@ class DynaLoraModel(LoraModel, DynaLoraMixin):
     def __init__(self,
                  model: PreTrainedModel,
                  peft_config: PeftConfig,
-                 adapter_name: str) -> None:
-        #peft_config.target_modules = "all-linear" # this would unfortunately inject LoRA on the random initialized classification layers too
-        if isinstance(model, RobertaForSequenceClassification):
-            peft_config.target_modules = target_modules=["key","query","value",
-                                                        "attention.output.dense",
-                                                        "intermediate.dense",
-                                                        "output.dense"]
-        else:
-            print("Dynalora is only supported for RobertaForSequenceClassification for now.")
-            exit(1)
-        # TODO: Define target modules for other models
+                 adapter_name: str = 'default') -> None:
         LoraModel.__init__(self, model, peft_config, adapter_name) # this would inject LoRA on the classification layers too
         DynaLoraMixin.__init__(self, adapter_name, peft_config)
 
@@ -197,7 +199,7 @@ class DinaLoraModel(LoraModel, DinaLoraMixin):
     def __init__(self,
                  model: PreTrainedModel,
                  peft_config: PeftConfig,
-                 adapter_name: str) -> None:
+                 adapter_name: str = 'default') -> None:
         LoraModel.__init__(self, model, peft_config, adapter_name)
         DinaLoraMixin.__init__(self, adapter_name, peft_config)
